+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to the DSH Api Client project will be documented in this fil
 * Some naming.
 * Improved documentation.
 * Changed type of guid to u16.
+* Improved error handling in client factory.
 
 ### Removed